replay-engine/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
redis==5.0.1
pydantic==2.5.0
pytest==7.4.3
pytest-asyncio==0.21.1
prometheus-client==0.19.0
PyYAML==6.0.1
gunicorn==21.2.0 
flask==3.0.0
requests